requirements.txt
google-generativeai
pypdf
python-docx
streamlit

